fix: fixing gemini request
