--- codigo-fonte/cultivo/apps/web/src/app.tsx
@@ -1,12 +1,0 @@
-import Assistente from "./components/assistente";
-
-function App() {
-  return (
-    <div>
-      <h1>Meu Projeto com Assistente Virtual</h1>
-      <Assistente />
-    </div>
-  );
-}
-
-export default App;
