+++ codigo-fonte/cultivo/packages/backend/convex/chat.ts
@@ -1,3 +1,4 @@
+import { GoogleGenAI } from "@google/genai";
 import { action, mutation } from "./_generated/server";
 import { v } from "convex/values";
 
@@ -7,6 +8,8 @@ const briefing = `
   Você é um assistente de IA que responde perguntas sobre a cultura do feijão.
   Responda em português, de forma clara e objetiva, e use markdown.
 `;
+
+const gemini = new GoogleGenAI({ apiKey: process.env.GEMINI_API_KEY });
 
 export const savePrompt = mutation({
   args: {
@@ -32,30 +35,28 @@ export const promptAssistant = action({
       throw new Error("A variável GEMINI_API_KEY não está configurada.");
     }
 
-    const response = await fetch(
-      "https://generativelanguage.googleapis.com/v1beta/models/gemini-1.5-flash-latest:generateContent?key=" +
-        GEMINI_API_KEY,
-      {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          contents: [
-            { role: "user", parts: [{ text: briefing }] },
-            { role: "user", parts: [{ text: prompt }] },
-          ],
-        }),
-      }
-    );
+    const model = "gemini-2.5-flash";
 
-    const data = await response.json();
+    const contents = [
+      { role: "user", parts: [{ text: briefing }] },
+      { role: "user", parts: [{ text: prompt }] },
+    ];
 
-    // Extração segura da resposta
-    const resposta =
-      data.candidates?.[0]?.content?.parts?.[0]?.text ||
-      "Não consegui gerar uma resposta.";
+    const response = await gemini.models.generateContent({
+      model,
+      contents,
+    });
 
-    return resposta;
+    const chatResponse = response.text || "Não consegui gerar uma resposta.";
+
+    // TODO: Salvar a conversa no banco de dados
+    // ctx.runMutation(api.chat.savePrompt, {
+    //   userId: null,
+    //   message: prompt,
+    //   response: chatResponse,
+    //   createdAt: Date.now(),
+    // });
+
+    return chatResponse;
   },
 });
