--- codigo-fonte/cultivo/packages/backend/tsconfig.json
@@ -1,25 +1,0 @@
-{
-	/* This TypeScript project config describes the environment that
-	 * Convex functions run in and is used to typecheck them.
-	 * You can modify it, but some settings required to use Convex.
-	 */
-	"compilerOptions": {
-		/* These settings are not required by Convex and can be modified. */
-		"allowJs": true,
-		"strict": true,
-		"moduleResolution": "Bundler",
-		"jsx": "react-jsx",
-		"skipLibCheck": true,
-		"allowSyntheticDefaultImports": true,
-
-		/* These compiler options are required by Convex */
-		"target": "ESNext",
-		"lib": ["ES2021", "dom"],
-		"forceConsistentCasingInFileNames": true,
-		"module": "ESNext",
-		"isolatedModules": true,
-		"noEmit": true
-	},
-	"include": ["./**/*"],
-	"exclude": ["./_generated"]
-}
